COM-3634: sophie's review
