+++ src/helpers/trainerMissions.js
@@ -1,11 +1,27 @@
 const get = require('lodash/get');
 const GCloudStorageHelper = require('./gCloudStorage');
 const UtilsHelper = require('./utils');
+const CourseSlotsHelper = require('./courseSlots');
+const StepsHelper = require('./steps');
 const { TRAINING_ORGANISATION_MANAGER, VENDOR_ADMIN, GENERATION, UPLOAD } = require('./constants');
 const Course = require('../models/Course');
 const TrainerMission = require('../models/TrainerMission');
 const TrainerMissionPdf = require('../data/pdf/trainerMission');
 
+const getFileName = (programName, trainerIdentity) =>
+  `ordre mission ${programName} ${UtilsHelper.formatIdentity(trainerIdentity, 'FL')}`;
+
+const uploadDocument = async (payload, courseIds, file, fileName, method, credentials) => {
+  const fileUploaded = await GCloudStorageHelper.uploadCourseFile({ fileName, file });
+
+  await TrainerMission.create({
+    ...payload,
+    courses: courseIds,
+    file: fileUploaded,
+    createdBy: credentials._id,
+    creationMethod: method,
+  });
+};
 exports.upload = async (payload, credentials) => {
   const courseIds = Array.isArray(payload.courses) ? payload.courses : [payload.courses];
   const course = await Course
@@ -17,18 +33,9 @@ exports.upload = async (payload, credentials) => {
     .lean();
 
   const programName = course.subProgram.program.name;
-  const trainerName = UtilsHelper.formatIdentity(course.trainer.identity, 'FL');
+  const fileName = getFileName(programName, course.trainer.identity);
 
-  const fileName = `ordre mission ${programName} ${trainerName}`;
-  const fileUploaded = await GCloudStorageHelper.uploadCourseFile({ fileName, file: payload.file });
-
-  await TrainerMission.create({
-    ...payload,
-    courses: courseIds,
-    file: fileUploaded,
-    createdBy: credentials._id,
-    creationMethod: UPLOAD,
-  });
+  return uploadDocument(payload, courseIds, payload.file, fileName, UPLOAD, credentials);
 };
 
 exports.list = async (query, credentials) => {
@@ -49,8 +56,7 @@ exports.list = async (query, credentials) => {
     .lean();
 };
 
-exports.generate = async (payload, credentials) => {
-  const courseIds = Array.isArray(payload.courses) ? payload.courses : [payload.courses];
+const formatData = async (courseIds, fee, credentials) => {
   const courses = await Course
     .find({ _id: { $in: courseIds } }, { hasCertifyingTest: 1, misc: 1, type: 1 })
     .populate({ path: 'companies', select: 'name' })
@@ -70,19 +76,26 @@ exports.generate = async (payload, credentials) => {
     trainerIdentity: trainer.identity,
     program: subProgram.program.name,
     slotsCount: slots.length + slotsToPlan.length,
-    liveDuration: UtilsHelper.computeLiveDuration(slots, slotsToPlan, subProgram.steps),
+    liveDuration: StepsHelper.computeLiveDuration(slots, slotsToPlan, subProgram.steps),
     groupCount: courses.length,
     companies: [...new Set(courses.map(c => UtilsHelper.formatName(c.companies)))].join(', '),
-    addressList: UtilsHelper.getAddressList(courses.map(c => c.slots).flat(), subProgram.steps),
-    dates: UtilsHelper.formatSlotDates(courses.map(c => c.slots).flat()),
+    addressList: CourseSlotsHelper.getAddressList(courses.map(c => c.slots).flat(), subProgram.steps),
+    dates: CourseSlotsHelper.formatSlotDates(courses.map(c => c.slots).flat()),
     slotsToPlan: courses.reduce((acc, course) => acc + course.slotsToPlan.length, 0),
     certification: courses.filter(c => c.hasCertifyingTest),
-    fee: payload.fee,
+    fee,
     createdBy: UtilsHelper.formatIdentity(credentials.identity, 'FL'),
   };
 
+  return data;
+};
+
+exports.generate = async (payload, credentials) => {
+  const courseIds = Array.isArray(payload.courses) ? payload.courses : [payload.courses];
+  const data = await formatData(courseIds, payload.fee, credentials);
+
   const pdf = await TrainerMissionPdf.getPdf(data);
-  const fileName = `ordre mission ${data.program} ${UtilsHelper.formatIdentity(data.trainerIdentity, 'FL')}`;
+  const fileName = getFileName(data.program, data.trainerIdentity);
   pdf.hapi = {
     fileName,
     headers: {
@@ -91,13 +104,5 @@ exports.generate = async (payload, credentials) => {
     },
   };
 
-  const fileUploaded = await GCloudStorageHelper.uploadCourseFile({ fileName, file: pdf });
-
-  await TrainerMission.create({
-    ...payload,
-    courses: courseIds,
-    file: fileUploaded,
-    createdBy: credentials._id,
-    creationMethod: GENERATION,
-  });
+  return uploadDocument(payload, courseIds, pdf, fileName, GENERATION, credentials);
 };
